Fix the way that the bridhe is dealt with
At some point in the past the way that modules are loaded has
gone away from this and the previous "t::Bridge" thing wasn't
working.

This may not be ideal but it gets it working again.

Will also update the ecosystem so that this one is used and not
the missing one in ingy's github